fix/Examples route tests
Changed:
1. Get examples test to a meaningful one
2. DB cleanup now works before and after every test

Test utils:
1. Test database gets dropped only if it exists

+++ backend/app/tests/utilsTests.py
@@ -4,7 +4,7 @@ Utils functions for tests.
 To be reused in other tests.
 """
 
-from sqlalchemy_utils.functions import drop_database as dropDatabase
+from sqlalchemy_utils.functions import drop_database as dropDatabase, database_exists
 from app.database.database import engine
 from app.database import models
 
@@ -15,7 +15,8 @@ def cleanupDB(start: bool):
 
   Used frequently to encapsulate tests as needed.
   """
-  dropDatabase(engine.url)
+  if database_exists(engine.url):
+    dropDatabase(engine.url)
 
   if start:
     models.Base.metadata.create_all(bind=engine)
